Create orders app, and move order-related models there
